fix bug datatype

+++ STK.Application/DTOs/RequisiteDto.cs
@@ -8,7 +8,7 @@ namespace STK.Application.DTOs
         public string OGRN { get; set; }
         public DateTime? DateCreation { get; set; }
         public string EstablishmentCreateName { get; set; }
-        public int? AuthorizedCapital { get; set; }
+        public decimal? AuthorizedCapital { get; set; }
         public int? AvgCountEmployee { get; set; }
         public string TypeOfCapital { get; set; }
     }
